microservices/requirements.txt
Flask==3.0.3
ollama==0.2.0
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2